moved Choices into choices.py
