+++ src/py42/services/detectionlists/departing_employee.py
@@ -1,9 +1,9 @@
 from datetime import datetime
 
+from py42.choices import Choices
 from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42NotFoundError
 from py42.exceptions import Py42UserNotOnListError
-from py42.sdk.queries import Choices
 from py42.services import BaseService
 from py42.services.detectionlists import _DetectionListFilters
 from py42.services.detectionlists import _PAGE_SIZE
